+++ helloworld/hello.go
@@ -1,4 +1,4 @@
-package main
+package helloworld
 
 import (
 	"fmt"
